First pass at refactoring the spec converter
This broadly splits the code into three subnamespaces:

- Grammar, which is aware of colorization but only that
- Spec, which is aware of Grammar and Markdown
- Converter, which is aware of Spec and Word

The Program class coordinates all of this.

There's a lot more refactoring to do, as well as avoiding public fields etc, but this is a starting point.

I've removed the capability of HTML output (to remove an unnecessary dependency) but haven't cleaned up all the command line parsing code that still accepts it.

I suspect the colorizer cache is unnecessary, and removing that and the "interactive mode" would simplify things further.